Update app.js
Error Checking for if the JSON and PDF don't have any keys or fields or if they're blank

+++ PdfAutofiller/app.js
@@ -22,25 +22,42 @@ app.use(express.static('public'))
 const port = 1337  //Host on port 1337
 
 app.post("/", (req, res) => {
-    const data = validateJSON(req.files[0].buffer.toString()) //Checks if the first file is JSON data or not
-    if (data) {
+   var stop = 0 //If this variable equals 1 it means that either the JSON or PDF file is empty
+
+    if (validateJSON(req.files[0].buffer.toString())) {
         var jsonData = req.files[0].buffer.toString()  //If JSON is the first file
         jsonData = JSON.parse(jsonData)
         b64 = req.files[1].buffer.toString('base64')
     }
-    else {
-        var jsonData = req.files[1].buffer.toString() //If PDF is the first file
-        jsonData = JSON.parse(jsonData)
-        b64 = req.files[0].buffer.toString('base64')
+    else if (validateJSON(req.files[1].buffer.toString())) {
+         var jsonData = req.files[1].buffer.toString() //If PDF is the first file
+         jsonData = JSON.parse(jsonData)
+         b64 = req.files[0].buffer.toString('base64')
     }
 
-    run(b64, jsonData)  //fill in PDF using the data from the 2 files
-        .then(PDF => {
-            console.log(PDF)
-            res.contentType("application/pdf")
-            res.send(PDF)
-            fs.unlinkSync('./asdhwbvjhsavd_filled.pdf')  //removes saved filled pdf
-        })
+    else { //If the JSON file is not provided
+         console.log("The JSON file is empty or not provided")
+         res.send("The JSON file is empty or not provided")
+         stop = 1
+         res.end()
+    }
+
+    if (b64.length == 0 && stop == 0) { //If the PDF file is empty
+        console.log("The PDF file is empty")
+        res.send("The PDF file is empty")
+        stop = 1
+        res.end()
+    }
+
+    if (stop == 0) {
+        run(b64, jsonData)  //fill in PDF using the data from the 2 files
+            .then(PDF => {
+                console.log(PDF)
+                res.contentType("application/pdf")
+                res.send(PDF)
+                fs.unlinkSync('./asdhwbvjhsavd_filled.pdf')  //removes saved filled pdf
+            })
+    }
 })
 
 app.get("/", (req, res) => {  //homepage path
@@ -128,41 +145,74 @@ function validateJSON(data) {
         return false
     }
 }
+
+function verify(fields, key, b64) {
+    var valid = 0
+    if (key.length == 0) {
+        console.log("Your json file has no values in it")
+        valid = 1
+    }
+
+    if (fields.length == 0) {
+        console.log("Your PDF file is not fillable")
+        valid = 1
+    }
+
+    if (b64.length == 0) {
+        console.log("Your PDF file is empty")
+        valid = 1
+    }
+
+    return valid
+}
+
 async function run(b64, jsonData) {
     let decodePdf = await base64.base64Decode(`${b64}`, 'asdhwbvjhsavd.pdf')  //decode base64 to PDF and saves it
     const pdf = await PDFDocument.load(fs.readFileSync('./asdhwbvjhsavd.pdf'))  //open decoded PDF
     const formPdf = pdf.getForm();  //get forms and fields PDF metadata
     const fields = formPdf.getFields();
     const key = Object.keys(jsonData)  //get keys from JSON
-    fields.forEach(field => {  //for every field
-        const type = field.constructor.name
-        const name = field.getName()
-        if (`${type}` == "PDFTextField") {  //fill in Text fields
-            if (key.includes(`${name}`)) {
-                field.setText(jsonData[name])
-            }
-        } else if (`${type}` == "PDFDropdown") {  //select from dropdown
-            if (key.includes(`${name}`)) {
-                field.select(jsonData[name])
-            }
-        } else if (`${type}` == "PDFCheckBox") {  //select checkboxes
-            if (key.includes(`${name}`)) {
-                if (jsonData[name] == 1) {
-                    field.check()
-                } else {
-                    field.uncheck()
+    const valid = verify(fields, key, b64)
+    var matches = 0
+    
+    if (valid == 0) {
+        fields.forEach(field => {  //for every field
+            const type = field.constructor.name
+            const name = field.getName()
+            if (`${type}` == "PDFTextField") {  //fill in Text fields
+                if (key.includes(`${name}`)) {
+                    matches += 1
+                    field.setText(jsonData[name])
+                }
+            } else if (`${type}` == "PDFDropdown") {  //select from dropdown
+                if (key.includes(`${name}`)) {
+                    matches += 1
+                    field.select(jsonData[name])
+                }
+            } else if (`${type}` == "PDFCheckBox") {  //select checkboxes
+                if (key.includes(`${name}`)) {
+                    matches += 1
+                    if (jsonData[name] == 1) {
+                        field.check()
+                    } else {
+                        field.uncheck()
+                    }
+                }
+            } else if (`${type}` == "PDFRadioGroup") {  //select radio button (button with only 1 allowed E.X. test question with answers A, B, C, D)
+                if (key.includes(`${name}`)) {
+                    matches += 1
+                    field.select(jsonData[name])
                 }
             }
-        } else if (`${type}` == "PDFRadioGroup") {  //select radio button (button with only 1 allowed E.X. test question with answers A, B, C, D)
-            if (key.includes(`${name}`)) {
-                field.select(jsonData[name])
-            }
-        }
-    })
-    formPdf.flatten();  //flattens the PDF (marks as Read-Only/Non-fillable/Filled/etc.)
+        })
+    }
+    console.log("You your json and PDF have",matches,"matching fields")
+    form_pdf.flatten();  //flattens the PDF (marks as Read-Only/Non-fillable/Filled/etc.)
     fs.writeFileSync('./asdhwbvjhsavd_filled.pdf', await pdf.save());  //save the pdf with a gibberish name to not overwrite any of the user's pdfs
     const filled_pdf = fs.readFileSync('./asdhwbvjhsavd_filled.pdf')
     fs.unlinkSync('./asdhwbvjhsavd.pdf')  //removes original unfilled pdf
+    let encodedPDF = await base64.base64Encode('./asdhwbvjhsavd_filled.pdf')  //encode into base64
+    console.log(typeof encodedPDF)
     return await filled_pdf
 }
